move the test utilities to the roman_datamodels directory as test_utils.py
